v0.4 - Object Tool Mk1
 -- Inputs --
Added a ObjectToolInput property to the ConfigData, as well as creating a new InputAction to handle it

 -- Player - Controller --
Implemented SwapToObjectTool, which simply calls SwapCurrentActiveTool to swap to the Object tool
 - Also added a Action Bind with the new input

 -- Player - Character --
Added a rotation snap size.  Implemented identically to grid snap size, with all functions (getter, change snap size, increment and decrement) and properties (CurrentRotationSnapValue and RotationSnapValues)
 - Also added a UI update call on BeginPlay
Added a function to return a pointer to the Player_Tools class spawned

 -- UI - Master --
Added two buttons and a text block for the rotation snap size, implemented identically to grid snap size
Added a function to get the player tools
 - Simply calls GetPT on the Player_Character
Added a ObjectState component (see below)

 -- UI - Object --
Implemented.
Added three TileView components and a button
 - TileView One displays available categories the player can select
 - TileView Two displays applicable sub categories
 - TileView Three displays all items that match the categories selected
 - The Button swaps between enabling RotationMode
Added a TMap of EObjectType/Array of Strings to denote the subcategories per category
Added a pointer back to the Master UI class
Added properties for the selected category and sub category, as well as pointers to the current button of each
On NativeConstruct, creates category buttons for each value of the enum
 - Added the macro ENUM_RANGE_BY_COUNT
Also creates a bind for RotationModeButton
Implemented UpdateObjectTileView
 - Starts by clearing the TileView, then stores all the row names of the ObjectDataTable in a TArray
 - Next, figures out what items should be shown. If no category is selected, show all items
 - If a category is selected, but a sub-category is not selected, show all of the category items only
 - Else both a Category and SubCategory are selected, so only show items of the sub-category
Implemented UpdateSubCategoryButtons
 - Starts by clearing the SubCategoryTileView, then finds the matching sub category index in the map.  For each sub category create a USubCategoryButtonData and add it to the list view
Implemented AddItemToObjectTileView
 - Called in UpdateObjectTileView, simply creates a new UObject to hold the item data and adds the object to the tile view
Implemented UpdateSelectedCategory and UpdateSelectedSubCategory
 - Update CurrentSelectedObjectType and updates the SubCatButtons, then update the ObjectTileView
 - UpdateSelectedSubCategory does the same but only updates the SelectedSubCategory
Implemented UpdateSelectedObject
 - Calls UpdateObjectMesh in the PlayerTools

 -- UI - Buttons --
Added buttons for the category and item tile views
 - Both consist of a image and a button, while the category button has an addition text box
 - On NativeConstruct, creates a OnReleased bind for the button (implemented in BP)
 - See Images v04-001 and v04-002

 -- UI - Object Data --
Holds the data for the three UObject classes used by each of the TileViews

 -- Data - ObjectData --
Implemented EObjectType, which are the available categories for items
Also Implemented FObjectData, which contains all the information about an item
As well as FSubCats, which is just a workaround for TArrays not being available in TMaps

 -- Player - Tools --
Now hides the BuildToolWidgetComponent on BeginPlay
Updated SwapTool to now hide the BuildToolWidgetComponent, unless the BuildTool is selected
Added StaticMeshComponent ObjectToolMeshComponent, used to display the mesh selected from the ItemListView in the UI
Updated Tick to now have a path for the ObjectTool
 - Starts by checking what mode they are in (Normal or Rotation)
 - If they are in Normal, moves the ObjectToolMeshComponent to a suitable snap location based on the selected grid snap size
 - If they are in Rotation Mode, snaps the ObjectToolMeshComponent rotation to the nearest RotatingSnap based on the selected rotation snap size and the mouse's location

// Object Tool Mk1 has been implemented successfully.  It is used by sorting through a Data Table of items based on selected categories, selecting an item from the output list and setting it in PlayerTools.  The player can then move and rotate it to where they want it placed, then release PrimaryTool to place it onto the World_BuildingLevel selected

 // It will need a Mk2, as currently the objects are simple static meshes and not actual classes.  Additionally, they don't interact with the money the player owns

 -- Editor - Icon Tool --
Added an icon tool, to simplify creating icons for the objects
 - Consists of two classes - a Editor Utility Widget and an Actor
 - Once opened, the EUW spawns the actor in the world.  Selecting a StaticMesh from the Content Drawer displays the mesh on the actor.  The user can then interact with the buttons to rotate and zoom in/around the model selected
 - The user can enter a custom name for the widget in the TextBox at the top. Then pressing the MakeIcon button will the create a Texture2D asset via RenderTargetCreateStaticTextureEditorOnly
 - Also outputs a message based on if the icon creation was successful or not
 - See Images v04-003 onwards