v0.5e - Build Tool Mk3.5
 -- General --
Improved commenting across the board

 -- BuildToolData --
Renamed the BuildingData files to this
Moved FBuildToolData from BuildToolDisplay to this class
Added FSelectableWallData, used to store information about selectable walls in a data table

 -- Player - BuildToolDisplay --
Updated GetDisplayWallsInUse to no longer check if the mesh is either WallMesh or HalfWallMesh
 - Now only checks if the mesh is not invalid
Updated GenerateNewBuildDisplay
 - Now checks if a mesh has been selected or that the player is in erase mode before drawing
CreateXWall and CreateYWall now set the SMC to use the SelectedMesh rather than the HalfWallMesh
 - Additionally, if the player is in erase mode and a mesh is already placed at a location, the mesh of the SMC is matched to the mesh found
Moved AddNewStaticMeshComponent to the Utility Functions section

 -- Player - Tools --
Updated the Build Tool snap range to 250 (a full wall size)
Implemented UpdateSelectedWall, which is used to update the selected mesh in the BuildToolDisplay

-- TestBox --
Removed

-- UI - Data --
Renamed UI_Player_ObjectData to UI_Data_Object to better define it as a object class
Added UI_Data_Build, to hold the UObject class for the UWallButtonData

-- UI - Build --
Added UI_Player_Build_WallButton, used by the player to select a wall from a list (see Images
Added a tile view component to UI_Player_Build, as well as a pointer to the WallDataTable and a pointer to the current selected mesh
On BeginPlay, gets all rows in the data table and adds a button that contains the wall data for each data table row
Implemented UpdateSelectedWall, which changes the selected wall to a new one.  If the selected wall matches the previous one, it is instead cleared.

-- World - BuildingLevel --
Implemented AddToBuildData, which adds a single new SMC to the BuildData array (rather than clearing what the array contains and starting from the begining)
Moved GetWallObjectMeshAtPosition and AddNewStaticMeshComponent to Utility Functions

 -- Icon Creation Tool --
Added two new rows, one for moving the object on the Y axis and one for moving the object on the Z axis (see Images/)

// Still need to do work on the build tool, but it is closer.  Need to figure out how to do refunding money on removal of wall (put price in the BuildData?  would need to have 2 cost properties on each build data which could be confusing / move data table? the UI needs to know the data table, maybe move to tools.  on removal search data table for matching mesh, but how memory intense is this?  which class does the searching?  does BuildingLevel do it and return the total cost?)