v0.4c - Build Tool Mk4.2
 -- BuildToolData --
Added a function to FWallData to update the ID to a new one

 -- UI - BuildWidget --
Now checks if X is less than 0.  If so, no text is set in the SizeText box, effectively hiding it

 -- UI - Player - Build --
Updated use of UpdateSelectedWall to UpdateSelectedItem

 -- World - BuildingLevel --
Added UpdateWallID, used to update an ID of a WallData based on a location and rotation
 - Simply finds the index of the wall in the BuildData array via the location, then changes the XWallData or YWallIndex ID based on the rotation

 -- Player - BuildToolDisplay --
Added a default argument to EndPos in GenerateBuildDisplay
Updated UpdateDisplayValidity to now use a int argument instead of a bool one.
 - Now checks if the input is 0, 1 or 2.  0 and 1 are identical to before but 2 now sets the colour to the hidden colour
Changed SelectedWallID to SelectedID
UpdateSubTool now cleares the selected ID and Mesh, alongside clearing the Build Display
Implemented GenerateWindowDisplay
 - Starts by moving the BuildToolDisplay to the StartPosition, followed by checking if there is a SMC in index 0 of SMCPool and that a mesh is set
 - If so, set SMCPool[0] to it if it hasn't already and move the SMC to the start location offsetting it by 125 on the x axis)

 -- Player - Tools --
Added two pointers, one to the last hit StaticMeshComponent and the other to it's original mesh
Renamed UpdateSelectedWall to UpdateSelectedItem, as well as adding a EBuildToolSubType argument
Added a path in OnTick if the BuildingTool is selected and it is in Window SubTool
 - Simply fires a trace, checking if the actor hit is a World_BuildingLevel
 - If it is, checks if it hit a StaticMeshComponent.  Then cast to the component and checks if the mesh is a valid mesh to change
 - If all correct, stores the mesh in LastHitSMC, stores the mesh in LastHitMesh then changes the mesh of the SMC to the chosen window mesh
 - If a valid static mesh component isn't hit, the BuildToolDisplay is set to invalid
Added sub tool path to OnReleased
 - First checks if there is a valid LastHitSMC.  Next, checks if it has been changed already.
 - If it has, update the wall ID in the BuildData of the mesh that has been changed, clear the LastHitSMC's and Mesh as the wall is already replaced and finally charge the player for the window

// Thats window mode in.  Need to update WallData to have all the windows in as well, so erase mode can be used on it.  Also should merge the two tables together