v0.6k - Object Spawn Update
Updated the spawning of new objects via the ObjectTool.  Also gave the ability to right-click to cancel placement of both Object and Build Tools

 -- World - BuildingLevel --
Added a separate static mesh pool for Object
Also added a new function to add new meshes to this pool (AddNewObjectComponent)
Updated the AddObjectToLevel function to now utilize the new ObjectPool

 -- UI - Object --
Removed the RotationMode button and it's OnReleased bind

 -- Tool - Parent --
Added SecondaryActionPressed and SecondaryActionReleased, works in a similar way to PrimaryActionPressed/Released

 -- Tool - Build --
Added override function SecondaryActionPressed.
 - When pressed, clears the current display and resets the placement

 -- Tool - Object --
Added override function SecondaryActionPressed.
 - When pressed, toggles rotation mode and clears the object's place location
Also added bool bPlacing, toggled when the primary action is pressed and released
When the PrimaryAction is first pressed, if an object has been selected the object is locked to it's position and rotation mode is entered.  When released, the object is placed into the world on that rotation.
Added a small rotation offset in RotationMode.  May need to be updated for each object