v0.7c - Object Tool Mk2.2
Added deleting, invalidation of objects, picking up placed objects

 -- World - BuildingLevel --
Removed the TArray of spawned objects, as this is no longer needed

 -- Tools - Parent --
Added an overridable DeleteAction function
Updated FireTraceToActor to now take in a second argument of ECollisionChannel to change the channel of the trace
Moved the RefundMultiplier from Player_Tool_Build to Parent

 -- Input and Player --
Added a new DeleteAction input, as well as corresponding functions in Player_Character and Player_Controller

  -- Tool - Object --
Added a InvalidMaterial and a TArray of SelectedObjectMaterials for handling the object's materials
Added a TArray of OverlappedActors, used to store any actors currently being overlapped
Added a bool used to denote if the location and rotation the mesh is currently in is valid

 Changed the CollisionProfileName of the SelectedObjectMeshComponent from BuildingTool to it's own new ObjectTool
Added two dynamic overlap functions, one for begin overlapping and one for ending overlaps
 - When an actor starts overlapping the SOMC, it is added to the OverlappedActors.  If the array is now greater than 0 (overlapping at least one object), then change the material of the SelectedObject to the InvalidMaterial.
 - When an actor stops overlapping, it is removed from the array.  If the array is now empty, then change the material of the SelectedObject back to it's original material
Added a check to see if the player doesn't have either an SelectedObject or a SelectedID.  If they don't, set RotationMode to false in PrimaryActionPressed
 - Also when a placed object is picked up, it is now hidden and casted to.  The SOMC's mesh is then set to the objects, then the rotation is matched and the materials are stored in the SelectedObjectMaterials array.
PrimaryActionReleased now checks to see if the object can be placed at the location.  If it can, the picked up object is visible again and the SOMC is reset and cleared
If the item cannot be placed at this location, stop rotating and placing.
Implemented DeleteAction
 - If the player has picked up a placed item, get it's ID and find it in the DataTable.  If it is found, the object is destroyed and the player is refunded

The object tool now has two new functions - deleting objects that have been placed and stopping players placing objects inside each other.

Deleting is simple, just checks if the player is holding a placed item.  If they are, it gets it's ID and finds it's matching row in the data table.  The Price property is then multiplied by the RefundAmount and added to the player's total money

Invalidating placement was a bit more tricky than i first thought.  Instead of using the picked up object for overlap queries, it is instead hidden and the SOMC is used instead.  This stops the need to have overlap events on every object.

Additionally, reseting the object's colour back to it's original form was tricky, due to the amount of paths the player could take (fail placement then stop grabbing, cancel grab while invalid ect).  This was sorted by reseting the SOMC whenever the object was cleared or placed.

Finally, a small bug was introduced when sometimes picking up an object and having it immediately start overlapping another object would cause it to enter RotationMode.  This was fixed by adding the !SelectedObjectMeshComponent->GetStaticMesh() check at the start of PrimaryActionPressed