v0.7d - Object Tool Mk2.3
Added object snapping, allowing players to place objects next to each other with ease

 -- World - BuildingLevel --
Replaced the void return with a Object_Parent pointer in AddObjectToLevel
 - Now returns the object spawned
Removed RemoveObjectFromLevel, unused

  -- Object - Parent --
Removed the ObjectTags array and GetObjectTags

 -- Data - ObjectData --
Added FObjectSnappingData, used to store the data of a snapping point
Also added a Class property to FObjectData, as well as a TArray of FObjectSnappingData for SnapLocations

 -- Tools - ObjectTool --
Updated the collision from QueryOnly to QueryAndPhysics for the SelectedObjectMeshComponent
Overlap events now only update the materials if the placement location is still valid
Implemented GetObjectDataRow, which simply returns the row data from the ObjectDataTable of a supplied ID
 - Updated all previous row finders with the new function
When an object is selected or picked up, the object keys and snap type is stored if valid
Renamed HitObject to SelectedObject
Implemented GetPlacementIsValid, which returns true or false if all validity checks are true

ToolTick - Snapping
Renamed MouseLocation to TargetLocations
Now checks what ObjectPlacementType the object is
If the object is a wall, now checks if the trace hit a wall.  If it did, simply moves it to the traces location.
If the object is a Floor or Object, checks if the hit object has a snap type.  If it does, check each snap, converting to world space, and see if the trace is within the snap distance
If they are, checks if the selected object's snap type matches.
If they do, heck if the current object contains a snap matching this point.  If they do, update the location and rotation.
 - Rotation is calculated by finding the dot product between the object's direction + the selected snap point.
 - Location is calculated by placing the objects at the same location then moving the new one on an offset on both the forward and right vectors

Fixed some issues with UpdateObjectMesh, now clears all data successfully

Just need to implement a stop to only calculate the dot product and offset once unless the snap changed